# external/interface/httpintf/swagger/api.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this application is to provide API over HTTP to toggler service,\nin which you can interact with the service in a programmatic way.",
    "title": "provides API on HTTP layer to the toggler service.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "version": "0.2.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/release-flags": {
      "get": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "feature",
          "flag"
        ],
        "summary": "List all the release flag that can be used to manage a feature rollout.",
        "operationId": "listReleaseFlags",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/listReleaseFlagResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "description": "This operation allows you to create a new release flag.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "feature",
          "flag"
        ],
        "summary": "Create a release flag that can be used for managing a feature rollout.",
        "operationId": "createReleaseFlag",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "flag": {
                  "$ref": "#/definitions/Flag"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/createReleaseFlagResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/release-flags/{flagID}": {
      "put": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "feature",
          "flag",
          "pilot"
        ],
        "summary": "Update a release flag.",
        "operationId": "updateReleaseFlag",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "FlagID",
            "description": "FlagID is the release flag id or the alias name.",
            "name": "flagID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "flag": {
                  "$ref": "#/definitions/Flag"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/updateReleaseFlagResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/release-flags/{flagID}/pilots": {
      "get": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "feature",
          "flag"
        ],
        "summary": "List all the release flag that can be used to manage a feature rollout.",
        "operationId": "listReleasePilots",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ReleaseFlagID",
            "description": "ReleaseFlagID is the release flag id or the alias name.",
            "name": "flagID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/listReleasePilotResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "AppToken": [
              "[]"
            ]
          }
        ],
        "description": "This operation allows you to create a new release flag.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "release",
          "feature",
          "flag"
        ],
        "summary": "Create a release flag that can be used for managing a feature rollout.",
        "operationId": "createReleasePilot",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ReleaseFlagID",
            "description": "ReleaseFlagID is the release flag id or the alias name.",
            "name": "flagID",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pilot": {
                  "$ref": "#/definitions/ReleasePilotView"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/createReleasePilotResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Error": {
      "description": "Error contains the details of the error",
      "type": "object",
      "properties": {
        "code": {
          "description": "The constant code of the error that can be used for localisation",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 401
        },
        "message": {
          "description": "The message that describe the error to the developer who do the integration.\nNot meant to be propagated to the end-user.\nThe Message may change in the future, it it helps readability,\nplease do not rely on the content in any way other than just reading it.",
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/external/interface/httpintf/httpapi"
    },
    "Flag": {
      "description": "Flag is the basic entity with properties that feature flag holds",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/domains/release"
    },
    "ReleasePilotView": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "ID represent the fact that this object will be persistent in the Subject",
          "type": "string"
        },
        "enrolled": {
          "description": "IsParticipating states that whether the ManualPilot for the given feature is enrolled, or blacklisted",
          "type": "boolean",
          "x-go-name": "Enrolled"
        },
        "external_id": {
          "description": "ExternalID is the unique id that connect links a pilot with the caller services.\nThe caller service is the service that use the release toggles for example and need A/B testing or Canary launch.",
          "type": "string",
          "x-go-name": "ExternalID"
        },
        "release_flag_id": {
          "description": "ReleasePilotID is the reference ID that can tell where this user record belongs to.",
          "type": "string",
          "x-go-name": "ReleasePilotID"
        }
      },
      "x-go-package": "github.com/toggler-io/toggler/external/interface/httpintf/httpapi"
    },
...
  },
  "responses": {
...
    "createReleaseFlagResponse": {
      "description": "CreateReleaseFlagResponse",
      "schema": {
        "type": "object",
        "properties": {
          "flag": {
            "$ref": "#/definitions/Flag"
          }
        }
      }
    },
    "createReleasePilotResponse": {
      "description": "CreateReleasePilotResponse",
      "schema": {
        "type": "object",
        "properties": {
          "pilot": {
            "$ref": "#/definitions/ReleasePilotView"
          }
        }
      }
    },
...
    "errorResponse": {
      "description": "ErrorResponse will contains a response about request that had some kind of problem.\nThe details will be included in the body.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
...
    "listReleaseFlagResponse": {
      "description": "ListReleaseFlagResponse",
      "schema": {
        "type": "object",
        "properties": {
          "flags": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Flag"
            },
            "x-go-name": "Flags"
          }
        }
      }
    },
    "listReleasePilotResponse": {
      "description": "ListReleasePilotResponse",
      "schema": {
        "type": "object",
        "properties": {
          "pilots": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/ReleasePilotView"
            },
            "x-go-name": "Pilots"
          }
        }
      }
    },
...
    "updateReleaseFlagResponse": {
      "description": "UpdateReleaseFlagResponse",
      "schema": {
        "type": "object",
        "properties": {
          "flag": {
            "$ref": "#/definitions/Flag"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "AppKey": {
      "type": "apiKey",
      "name": "X-APP-KEY",
      "in": "header"
    },
    "AppToken": {
      "type": "apiKey",
      "name": "X-APP-TOKEN",
      "in": "header"
    }
  },
  "security": [
    {
      "AppKey": [
        "[]"
      ]
    }
  ]
}